requirements.txt
django==3.1.7
gunicorn==20.1.0
redis==4.3.1
celery==4.4.5
graphene-django==2.15.0
django-graphql-jwt==0.3.0
sgmllib3k==1.0.0
feedparser==5.2.1
requests==2.22.0
pyrfc3339==1.1
psycopg2-binary>=2.8
pyjwt==1.7.0
